renamed unit test files and tags to consistently use sub_level_error
