+++ test/catch2/sub_level_error/unit/test_sub_level_error_is_valid_sub_level_error.cpp
@@ -11,7 +11,7 @@
 #include "../../wrappers/connection_wrapper.h"
 
 /*
- * [is_valid_sub_level_error]: test_is_valid_sub_level_error.c
+ * [sub_level_error_is_valid_sub_level_error]: test_sub_level_error_is_valid_sub_level_error.c
  * This unit test file tests that the helper function __wt_is_valid_sub_level_error correctly
  * validates sub level error codes, which range between -32000 and -32199 inclusive.
  */
@@ -19,7 +19,7 @@
 TEST_CASE(
   "Test that helper function __wt_is_valid_sub_level_error validates sub level error codes "
   "correctly",
-  "[is_valid_sub_level_error],[sub_level_error]")
+  "[sub_level_error_is_valid_sub_level_error],[sub_level_error]")
 {
 
     /* Ensure no normal error codes are mistaken for sub level error codes. */
